requirements.txt
pycryptodome==3.20.0
Faker==25.2.0
matplotlib>=3.7.0
